feat: big view refactor, re-do logins
A lot was going on here at the same time and I didn't feel like backing out and separating it all.

First, templates are now loaded as an inheritance tree so register.html and login.html can share the init function in auth.html.

Second, DeleteChallenge wasn't acting right with the CreatedAt not being passed in.

Third, I moved passing the challenge ID from a cookie to a JSON field.